requirements.txt
PyYAML
pyqtgraph
PyQt5
numpy

